+++ watertap/examples/flowsheets/case_studies/seawater_RO_desalination/seawater_RO_desalination.py
@@ -44,7 +44,7 @@ from watertap.unit_models.reverse_osmosis_0D import (
     PressureChangeType,
 )
 from watertap.unit_models.pressure_exchanger import PressureExchanger
-from watertap.unit_models.pump_isothermal import Pump
+from watertap.unit_models.pressure_changer import Pump
 from watertap.core.util.initialization import assert_degrees_of_freedom
 
 from watertap.core.wt_database import Database
