+++ watertap/unit_models/pressure_changer.py
@@ -12,6 +12,7 @@
 ###############################################################################
 
 from pyomo.environ import Constraint
+from pyomo.common.config import In
 
 # Import IDAES cores
 from idaes.generic_models.unit_models.pressure_changer import PumpData
@@ -53,10 +54,11 @@ class PumpIsothermalData(PumpData):
 @declare_process_block_class("EnergyRecoveryDevice")
 class EnergyRecoveryDeviceData(PumpIsothermalData):
     """
-    Turbine-type Isothermal energy recovery device
+    Turbine-type isothermal energy recovery device
     """
+
     # switch compressor to False
     CONFIG = PumpIsothermalData.CONFIG()
-    CONFIG.compressor = False
     CONFIG.get("compressor")._default = False
     CONFIG.get("compressor")._domain = In([False])
+    CONFIG.compressor = False
